fix : no more warnings

+++ gui/src/raylib/debug/draw.cpp
@@ -19,11 +19,11 @@ namespace Zappy {
     void DebugMode::drawAnimations(void) {
         raylib::Color white = raylib::Color::White();
         raylib::Color green = raylib::Color::Green();
-    
+
         raylib::Vector2 startPos(1400, 100);
         white.DrawText("Animations: (I, K)", startPos.x, startPos.y - 70, 60);
         for (size_t i = 0; i < _animations.size(); i++) {
-            if (i == _animIndex) {
+            if ((int)i == _animIndex) {
                 green.DrawText("- " + std::string(_animations[i].name), startPos.x, startPos.y, 30);
             } else {
                 white.DrawText(std::string(_animations[i].name), startPos.x, startPos.y, 30);
