+++ gui/src/raylib/drawTiles.cpp
@@ -13,7 +13,7 @@ namespace Zappy
     {
         _rockModel.UpdateAnimation(_rockAnimations[_rockAnimationIndex], _rockAnimationFrame);
         _rockAnimationFrame++;
-        if (_rockAnimationFrame >= _rockAnimations[_rockAnimationIndex].frameCount) {
+        if (static_cast<int>(_rockAnimationFrame) >= _rockAnimations[_rockAnimationIndex].frameCount) {
             _rockAnimationIndex = rand() % 9 + 3;
             _rockAnimationFrame = 0;
         }
